backend/pyproject.toml
[tool.poetry]
name = "lwca"
version = "1.0.0"
description = ""
authors = ["Nicolas Kerdelhué"]

[tool.poetry.dependencies]
python = "3.8.8"
flask = "2.2.0"
uwsgi = "2.0.18"
flask-sqlalchemy = "3.0.3"
celery = "4.4.2"
flask-migrate = "2.5.3"
sqlalchemy-utils = "0.36.3"
cryptography = "2.8"
redis = "3.4.1"
flask-shell-ipython = "0.4.1"
psycopg2 = "2.8.4"
boto3 = "1.16.44"
localstack-client = "1.10"
flask-cors = "^4.0.1"
flasgger = "^0.9.7.1"
gitpython = "^3.1.43"
Werkzeug = "2.3.7"
sqlalchemy = "^2.0.31"
langchain-community = "^0.2.7"
flask-bcrypt = "^1.0.1"
flask-jwt-extended = "^4.6.0"
flask-socketio = "^5.3.6"

[tool.poetry.dev-dependencies]
pytest = "5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
